# feat: add db query latency metrics