--- a/lib/db/client-applications-store.js
+++ b/lib/db/client-applications-store.js
@@ -1,6 +1,9 @@
 /* eslint camelcase:off */
 'use strict';
 
+const metricsHelper = require('./metrics-helper');
+const { DB_TIME } = require('../events');
+
 const COLUMNS = [
     'app_name',
     'created_at',
@@ -35,15 +38,22 @@ const remapRow = (input, old = {}) => ({
 });
 
 class ClientApplicationsDb {
-    constructor(db) {
+    constructor(db, eventBus) {
         this.db = db;
+        this.eventBus = eventBus;
     }
 
     updateRow(details, prev) {
         details.updatedAt = 'now()';
         return this.db(TABLE)
             .where('app_name', details.appName)
-            .update(remapRow(details, prev));
+            .update(remapRow(details, prev))
+            .then(
+                metricsHelper.wrapTimer(this.eventBus, DB_TIME, {
+                    store: 'applications',
+                    action: 'updateRow',
+                })
+            );
     }
 
     insertNewRow(details) {
@@ -63,7 +73,13 @@ class ClientApplicationsDb {
                 } else {
                     return this.insertNewRow(data);
                 }
-            });
+            })
+            .then(
+                metricsHelper.wrapTimer(this.eventBus, DB_TIME, {
+                    store: 'applications',
+                    action: 'upsert',
+                })
+            );
     }
 
     getAll() {
@@ -71,7 +87,13 @@ class ClientApplicationsDb {
             .select(COLUMNS)
             .from(TABLE)
             .orderBy('app_name', 'asc')
-            .map(mapRow);
+            .map(mapRow)
+            .then(
+                metricsHelper.wrapTimer(this.eventBus, DB_TIME, {
+                    store: 'applications',
+                    action: 'getAll',
+                })
+            );
     }
 
     getApplication(appName) {