--- a/lib/db/client-instance-store.js
+++ b/lib/db/client-instance-store.js
@@ -1,6 +1,9 @@
 /* eslint camelcase: "off" */
 'use strict';
 
+const metricsHelper = require('./metrics-helper');
+const { DB_TIME } = require('../events');
+
 const COLUMNS = [
     'app_name',
     'instance_id',
@@ -23,8 +26,9 @@ const mapRow = row => ({
 });
 
 class ClientInstanceStore {
-    constructor(db, getLogger) {
+    constructor(db, eventBus, getLogger) {
         this.db = db;
+        this.eventBus = eventBus;
         this.logger = getLogger('client-instance-store.js');
         const clearer = () => this._removeInstancesOlderThanTwoDays();
         setTimeout(clearer, 10).unref();
@@ -72,7 +76,13 @@ class ClientInstanceStore {
                 } else {
                     return this.insertNewRow(details);
                 }
-            });
+            })
+            .then(
+                metricsHelper.wrapTimer(this.eventBus, DB_TIME, {
+                    store: 'instance',
+                    action: 'insert',
+                })
+            );
     }
 
     getAll() {
@@ -80,7 +90,13 @@ class ClientInstanceStore {
             .select(COLUMNS)
             .from(TABLE)
             .orderBy('last_seen', 'desc')
-            .map(mapRow);
+            .map(mapRow)
+            .then(
+                metricsHelper.wrapTimer(this.eventBus, DB_TIME, {
+                    store: 'instance',
+                    action: 'getAll',
+                })
+            );
     }
 
     getByAppName(appName) {