--- a/lib/db/client-metrics-store.js
+++ b/lib/db/client-metrics-store.js
@@ -1,14 +1,17 @@
 'use strict';
 
 const { EventEmitter } = require('events');
+const metricsHelper = require('./metrics-helper');
+const { DB_TIME } = require('../events');
 
 const TEN_SECONDS = 10 * 1000;
 
 class ClientMetricsStore extends EventEmitter {
-    constructor(metricsDb, getLogger, pollInterval = TEN_SECONDS) {
+    constructor(metricsDb, eventBus, getLogger, pollInterval = TEN_SECONDS) {
         super();
         this.logger = getLogger('client-metrics-store.js');
         this.metricsDb = metricsDb;
+        this.eventBus = eventBus;
         this.highestIdSeen = 0;
 
         this._init(pollInterval);
@@ -45,7 +48,12 @@ class ClientMetricsStore extends EventEmitter {
 
     // Insert new client metrics
     insert(metrics) {
-        return this.metricsDb.insert(metrics);
+        return this.metricsDb.insert(metrics).then(
+            metricsHelper.wrapTimer(this.eventBus, DB_TIME, {
+                store: 'metrics',
+                action: 'insert',
+            })
+        );
     }
 
     destroy() {