--- a/lib/db/client-metrics-store.test.js
+++ b/lib/db/client-metrics-store.test.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const test = require('ava');
+const { EventEmitter } = require('events');
 const ClientMetricStore = require('./client-metrics-store');
 const lolex = require('lolex');
 const getLogger = require('../../test/fixtures/no-logger');
@@ -24,8 +25,8 @@ function getMockDb() {
 
 test.cb('should call database on startup', t => {
     const mock = getMockDb();
-
-    const store = new ClientMetricStore(mock, getLogger);
+    const ee = new EventEmitter();
+    const store = new ClientMetricStore(mock, ee, getLogger);
 
     t.plan(2);
 
@@ -42,8 +43,8 @@ test.cb('should start poller even if inital database fetch fails', t => {
     const clock = lolex.install();
     const mock = getMockDb();
     mock.getMetricsLastHour = () => Promise.reject('oops');
-
-    const store = new ClientMetricStore(mock, getLogger, 100);
+    const ee = new EventEmitter();
+    const store = new ClientMetricStore(mock, ee, getLogger, 100);
 
     const metrics = [];
     store.on('metrics', m => metrics.push(m));
@@ -64,7 +65,8 @@ test.cb('should poll for updates', t => {
     const clock = lolex.install();
 
     const mock = getMockDb();
-    const store = new ClientMetricStore(mock, getLogger, 100);
+    const ee = new EventEmitter();
+    const store = new ClientMetricStore(mock, ee, getLogger, 100);
 
     const metrics = [];
     store.on('metrics', m => metrics.push(m));