--- a/lib/db/index.js
+++ b/lib/db/index.js
@@ -9,7 +9,7 @@ const ClientMetricsDb = require('./client-metrics-db');
 const ClientMetricsStore = require('./client-metrics-store');
 const ClientApplicationsStore = require('./client-applications-store');
 
-module.exports.createStores = config => {
+module.exports.createStores = (config, eventBus) => {
     const getLogger = config.getLogger;
     const db = createDb(config);
     const eventStore = new EventStore(db, getLogger);
@@ -20,8 +20,16 @@ module.exports.createStores = config => {
         eventStore,
         featureToggleStore: new FeatureToggleStore(db, eventStore, getLogger),
         strategyStore: new StrategyStore(db, eventStore, getLogger),
-        clientApplicationsStore: new ClientApplicationsStore(db, getLogger),
-        clientInstanceStore: new ClientInstanceStore(db, getLogger),
-        clientMetricsStore: new ClientMetricsStore(clientMetricsDb, getLogger),
+        clientApplicationsStore: new ClientApplicationsStore(
+            db,
+            eventBus,
+            getLogger
+        ),
+        clientInstanceStore: new ClientInstanceStore(db, eventBus, getLogger),
+        clientMetricsStore: new ClientMetricsStore(
+            clientMetricsDb,
+            eventBus,
+            getLogger
+        ),
     };
 };