--- a/lib/events.js
+++ b/lib/events.js
@@ -2,4 +2,5 @@
 
 module.exports = {
     REQUEST_TIME: 'request_time',
+    DB_TIME: 'db_time',
 };