--- a/lib/metrics.js
+++ b/lib/metrics.js
@@ -30,6 +30,12 @@ exports.startMonitoring = (
         labelNames: ['path', 'method', 'status'],
         percentiles: [0.1, 0.5, 0.9, 0.99],
     });
+    const dbDuration = new client.Summary({
+        name: 'db_query_duration_milliseconds',
+        help: 'DB query duration time',
+        labelNames: ['store', 'action'],
+        percentiles: [0.1, 0.5, 0.9, 0.99],
+    });
     const featureToggleUpdateTotal = new client.Counter({
         name: 'feature_toggle_update_total',
         help: 'Number of times a toggle has  been updated',
@@ -43,6 +49,10 @@ exports.startMonitoring = (
 
     eventBus.on(events.REQUEST_TIME, ({ path, method, time, statusCode }) => {
         requestDuration.labels(path, method, statusCode).observe(time);
+    });
+
+    eventBus.on(events.DB_TIME, ({ store, action, time }) => {
+        dbDuration.labels(store, action).observe(time);
     });
 
     eventStore.on(FEATURE_CREATED, ({ data }) => {