--- a/lib/metrics.js
+++ b/lib/metrics.js
@@ -31,7 +31,7 @@ exports.startMonitoring = (
         percentiles: [0.1, 0.5, 0.9, 0.99],
     });
     const dbDuration = new client.Summary({
-        name: 'db_query_duration_milliseconds',
+        name: 'db_query_duration_seconds',
         help: 'DB query duration time',
         labelNames: ['store', 'action'],
         percentiles: [0.1, 0.5, 0.9, 0.99],