--- a/lib/metrics.test.js
+++ b/lib/metrics.test.js
@@ -5,7 +5,7 @@ const { EventEmitter } = require('events');
 const eventBus = new EventEmitter();
 const eventStore = new EventEmitter();
 const clientMetricsStore = new EventEmitter();
-const { REQUEST_TIME } = require('./events');
+const { REQUEST_TIME, DB_TIME } = require('./events');
 const { FEATURE_UPDATED } = require('./event-type');
 const { startMonitoring } = require('./metrics');
 const { register: prometheusRegister } = require('prom-client');
@@ -54,5 +54,19 @@ test('should collect metrics for client metric reports', t => {
     t.regex(
         metrics,
         /feature_toggle_usage_total{toggle="TestToggle",active="true"} 10\nfeature_toggle_usage_total{toggle="TestToggle",active="false"} 5/
+    );
+});
+
+test('should collect metrics for db query timings', t => {
+    eventBus.emit(DB_TIME, {
+        store: 'foo',
+        action: 'bar',
+        time: 1337,
+    });
+
+    const metrics = prometheusRegister.metrics();
+    t.regex(
+        metrics,
+        /db_query_duration_milliseconds{quantile="0\.99",store="foo",action="bar"} 1337/
     );
 });