--- a/lib/metrics.test.js
+++ b/lib/metrics.test.js
@@ -61,12 +61,12 @@ test('should collect metrics for db query timings', t => {
     eventBus.emit(DB_TIME, {
         store: 'foo',
         action: 'bar',
-        time: 1337,
+        time: 0.1337,
     });
 
     const metrics = prometheusRegister.metrics();
     t.regex(
         metrics,
-        /db_query_duration_milliseconds{quantile="0\.99",store="foo",action="bar"} 1337/
+        /db_query_duration_seconds{quantile="0\.99",store="foo",action="bar"} 0.1337/
     );
 });