--- a/lib/server-impl.js
+++ b/lib/server-impl.js
@@ -16,8 +16,8 @@ const { addEventHook } = require('./event-hook');
 async function createApp(options) {
     // Database dependencies (stateful)
     const logger = options.getLogger('server-impl.js');
-    const stores = createStores(options);
     const eventBus = new EventEmitter();
+    const stores = createStores(options, eventBus);
 
     const config = Object.assign(
         {