--- a/lib/timer.test.js
+++ b/lib/timer.test.js
@@ -12,11 +12,11 @@ function timeout(fn, ms) {
     );
 }
 
-test('should calculate the correct time in milliseconds', t => {
-    t.is(timer.milliseconds([1, 0]), 1000);
-    t.is(timer.milliseconds([0, 1e6]), 1);
-    t.is(timer.milliseconds([1, 1e6]), 1001);
-    t.is(timer.milliseconds([1, 552]), 1000.000552);
+test('should calculate the correct time in seconds', t => {
+    t.is(timer.seconds([1, 0]), 1);
+    t.is(timer.seconds([0, 1e6]), 0.001);
+    t.is(timer.seconds([1, 1e6]), 1.001);
+    t.is(timer.seconds([1, 552]), 1.000000552);
 });
 
 test('timer should track the time', async t => {
@@ -25,7 +25,7 @@ test('timer should track the time', async t => {
     await timeout(() => {
         diff = tt();
     }, 20);
-    if (diff > 19 && diff < 50) {
+    if (diff > 0.019 && diff < 0.05) {
         return t.pass();
     }
     t.fail();