**What this PR does**
This PR adds the `db_query_duration_milliseconds` metric to track the database query latency.

**Why do we need this?**
I want to monitor the timing of the DB queries to see if high http response times correlate with slow database queries. Without these metrics i can not tell whether the node process, the network to the DB or the DB itself is slow.

**Additional Notes**
While connecting more teams with the feature toggle we've seen latency issues on the following API paths:

Request volume on the Feature Toggle API Server: ~25 requests per second at peak times.

| call | 90th percentile | 99th percentile |
|---|---|---|
| POST /api/client/register  | 10ms -> 21ms  | 88ms -> 186ms  |
| POST /api/client/metrics  |  9ms -> 20ms| 100ms -> 120ms  |
| GET /api/client/features | 4ms -> 10ms | 25ms -> 210ms (!) |

We've implemented a simple in-memory cache for `GET /api/client/features.*` to resolve the latency issue.

Unfortunately, i cant get deep insights into the (managed) Database about query performance.