@ivarconr 2019-08-04
Metrics all the things :fire: 

Thanks, this looks good to me. Getting better metrics is always good. Can't see any big issue.